// src/domain/generated/contentful.d.ts
export interface IAdventcalendarFields {
  /** Identifier */
  identifier: string;

  /** Welcome Message */
  welcomeMessage?: string | undefined;

  /** Appearance */
  appearance?: "lebkuchenhaus" | "tiles" | undefined;

  /** Entries */
  entries?: ICalendarEntry[] | undefined;
}
// ...
export interface ICalendarEntryFields {
  /** Day */
  day?:
    | "1"
    | "2"
    | "3"
    | "4"
    | "5"
    | "6"
    | "7"
    | "8"
    | "9"
    | "10"
    | "11"
    | "12"
    | "13"
    | "14"
    | "15"
    | "16"
    | "17"
    | "18"
    | "19"
    | "20"
    | "21"
    | "22"
    | "23"
    | "24"
    | undefined;

  /** Description */
  text?: string | undefined;

  /** Image */
  image: Asset;
}

export interface ICalendarEntry extends Entry<ICalendarEntryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "calendarEntry";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}